requirements.txt
psycopg2
pandas
google-cloud-bigquery